+++ services/accounts/accountOperations.ts
@@ -11,7 +11,7 @@ import { UI_CONSTANTS } from "~/constants/ui"
 import {
   ensureDefaultApiTokenForAccount,
   generateDefaultTokenRequest,
-} from "~/services/accountKeyAutoProvisioning/ensureDefaultToken"
+} from "~/services/accounts/accountKeyAutoProvisioning/ensureDefaultToken"
 import { accountStorage } from "~/services/accounts/accountStorage"
 import { getApiService } from "~/services/apiService"
 import { autoDetectSmart } from "~/services/autoDetectService"
